animation added

+++ ModelViewer/Shaders/model.vert
@@ -2,7 +2,7 @@
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec3 aNormal;
 layout (location = 2) in vec2 aTexCoords;
-layout (location = 3) in vec4 BoneIDs;
+layout (location = 3) in ivec4 BoneIDs;
 layout (location = 4) in vec4 Weights;
 
 out vec2 TexCoords;
@@ -18,10 +18,10 @@ uniform mat4 gBones[100];
 void main()
 {
 
-    mat4 BoneTransform = gBones[int(BoneIDs[0])] * Weights[0];
-    BoneTransform += gBones[int(BoneIDs[1])] * Weights[1];
-    BoneTransform += gBones[int(BoneIDs[2])] * Weights[2];
-    BoneTransform += gBones[int(BoneIDs[3])] * Weights[3];
+    mat4 BoneTransform = gBones[BoneIDs[0]] * Weights[0];
+    BoneTransform += gBones[BoneIDs[1]] * Weights[1];
+    BoneTransform += gBones[BoneIDs[2]] * Weights[2];
+    BoneTransform += gBones[BoneIDs[3]] * Weights[3];
 
     vec4 pPos = BoneTransform * vec4(aPos, 1.0);
     vec4 pNormal = BoneTransform * vec4(aNormal, 1.0);
